Se agrega filtro por ingredientes

+++ app/src/main/java/com/somadtech/mrsushi/activities/CartActivity.java
@@ -169,10 +169,10 @@ public class CartActivity extends AppCompatActivity implements NavigationView.On
         return true;
     }
 
-    long getCategoryId(MenuItem item){
+    int getCategoryId(MenuItem item){
         String category_slug = item.getTitleCondensed().toString();
         if (!category_slug.equals("")) {
-            long category_id = mDbHelper.getCategoryId(category_slug);
+            int category_id = mDbHelper.getCategoryId(category_slug);
             return category_id;
         } else {
             return 1;
